refactor: extract activity summary code to make it a reusable component
