+++ src/activity-feed/components/activity/__tests__/index.test.tsx
@@ -8,13 +8,20 @@ import { activityByIdState } from "../../../selectors";
 
 jest.mock("recoil");
 jest.mock("../../../selectors");
+jest.mock("../../../../common/components/activity-summary", () => ({
+  ActivitySummary: ({ activity }: any) => (
+    <div data-component="ActivitySummary">
+      {JSON.stringify(activity, null, 2)}
+    </div>
+  ),
+}));
 
 describe("activity-feed/components/activity", () => {
   describe("Activity", () => {
     describe("#render", () => {
-      it("should render an anwsered outbound call correctly", () => {
+      it("should render the given activity correctly", () => {
         (useRecoilValue as jest.Mock).mockReturnValue({
-          id: 7834,
+          id: 42,
           created_at: "2018-04-19T09:38:41.000Z",
           direction: "outbound",
           from: "Pierre-Baptiste Béchu",
@@ -23,96 +30,6 @@ describe("activity-feed/components/activity", () => {
           duration: "120",
           is_archived: false,
           call_type: "answered",
-        });
-        const { container } = render(<Activity id={42} />);
-
-        expect(activityByIdState).toHaveBeenCalledWith(42);
-        expect(container).toMatchSnapshot();
-      });
-
-      it("should render a missed outbound call correctly", () => {
-        (useRecoilValue as jest.Mock).mockReturnValue({
-          id: 7834,
-          created_at: "2018-04-19T09:38:41.000Z",
-          direction: "outbound",
-          from: "Pierre-Baptiste Béchu",
-          to: "06 46 62 12 33",
-          via: "NYC Office",
-          duration: "120",
-          is_archived: false,
-          call_type: "missed",
-        });
-        const { container } = render(<Activity id={42} />);
-
-        expect(activityByIdState).toHaveBeenCalledWith(42);
-        expect(container).toMatchSnapshot();
-      });
-
-      it("should render an anwsered inbound call correctly", () => {
-        (useRecoilValue as jest.Mock).mockReturnValue({
-          id: 7834,
-          created_at: "2018-04-19T09:38:41.000Z",
-          direction: "inbound",
-          from: "Pierre-Baptiste Béchu",
-          to: "06 46 62 12 33",
-          via: "NYC Office",
-          duration: "120",
-          is_archived: false,
-          call_type: "answered",
-        });
-        const { container } = render(<Activity id={42} />);
-
-        expect(activityByIdState).toHaveBeenCalledWith(42);
-        expect(container).toMatchSnapshot();
-      });
-
-      it("should render a missed inbound call correctly", () => {
-        (useRecoilValue as jest.Mock).mockReturnValue({
-          id: 7834,
-          created_at: "2018-04-19T09:38:41.000Z",
-          direction: "inbound",
-          from: "Pierre-Baptiste Béchu",
-          to: "06 46 62 12 33",
-          via: "NYC Office",
-          duration: "120",
-          is_archived: false,
-          call_type: "missed",
-        });
-        const { container } = render(<Activity id={42} />);
-
-        expect(activityByIdState).toHaveBeenCalledWith(42);
-        expect(container).toMatchSnapshot();
-      });
-
-      it("should render a missed outbound call with voicemail correctly", () => {
-        (useRecoilValue as jest.Mock).mockReturnValue({
-          id: 7834,
-          created_at: "2018-04-19T09:38:41.000Z",
-          direction: "outbound",
-          from: "Pierre-Baptiste Béchu",
-          to: "06 46 62 12 33",
-          via: "NYC Office",
-          duration: "120",
-          is_archived: false,
-          call_type: "voicemail",
-        });
-        const { container } = render(<Activity id={42} />);
-
-        expect(activityByIdState).toHaveBeenCalledWith(42);
-        expect(container).toMatchSnapshot();
-      });
-
-      it("should render a missed inbound call with voicemail correctly", () => {
-        (useRecoilValue as jest.Mock).mockReturnValue({
-          id: 7834,
-          created_at: "2018-04-19T09:38:41.000Z",
-          direction: "inbound",
-          from: "Pierre-Baptiste Béchu",
-          to: "06 46 62 12 33",
-          via: "NYC Office",
-          duration: "120",
-          is_archived: false,
-          call_type: "voicemail",
         });
         const { container } = render(<Activity id={42} />);
