+++ src/common/theme/index.ts
@@ -72,8 +72,12 @@ export const theme = {
         COLOR: ABSTRACT.COLORS.ERROR,
       },
       ACTIVITY: {
-        INNER_SPACING: PRIMITIVE.SPACING.S,
         BORDER_COLOR: ABSTRACT.COLORS.MEDIUM_BORDER,
+      },
+    },
+    COMMON: {
+      ACTIVITY_SUMMARY: {
+        INNER_SPACING: PRIMITIVE.SPACING.S,
         PHONE_ICON_COLOR: ABSTRACT.COLORS.MEDIUM_ICON,
         MISSED_COLOR: ABSTRACT.COLORS.ERROR,
         ANSWERED_COLOR: ABSTRACT.COLORS.SUCCESS,
